feat: move dynamic lists to urql
