+++ lib/api/galleries/media-policy.ts
@@ -1,5 +1,5 @@
 import { graphql } from "@/gql/gql";
-import queryAPI from "@/lib/api/client/query";
+import queryAPI from "@/lib/api/client/server";
 import { getSiteFromLocale } from "@/lib/helpers/site";
 
 export const getMediaPolicyPage = async (locale: string) => {
