+++ lib/utils.js
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import GlobalDataContext from "@/contexts/GlobalData";
-import { useDataList } from "@/api/entries";
+import useDataList from "@/hooks/useDataList";
 import { useReleases } from "@/lib/api/noirlabReleases";
 import { useFilterParams } from "@/contexts/FilterParams";
 import { fallbackLng } from "./i18n/settings";
@@ -65,8 +65,9 @@ export const useList = ({
   const pathName = usePathname();
   const { params } = useFilterParams();
   const site = getSiteString(pathName);
-  const adjustedLimit = adjustLimit(limit, params.page, showsFeatured);
-  const offset = getOffset(limit, params.page, showsFeatured);
+  const page = parseInt(params.page);
+  const adjustedLimit = adjustLimit(limit, page, showsFeatured);
+  const offset = getOffset(adjustedLimit, page, showsFeatured);
   const inReverse = params.sort === "descending";
   const search = params.search ? `"${params.search}"` : null;
   const categories = useGlobalData("categories");
@@ -95,7 +96,7 @@ export const useList = ({
   if (entriesWithNoirlabReleases && results?.data) {
     const { entries, isLoading, isError } = entriesWithNoirlabReleases;
     results.data.entries = entries;
-    results.isLoading = isLoading;
+    results.isLoading = results.isLoading || isLoading;
     results.isError = isError;
   }
