+++ services/craft/globals/contactForm.ts
@@ -1,6 +1,6 @@
 "server-only";
 import { graphql } from "@/gql";
-import queryAPI from "@/lib/api/client/query";
+import queryAPI from "@/lib/api/client/server";
 import tags from "@/lib/api/client/tags";
 import { getSiteFromLocale } from "@/lib/helpers/site";
 import { getLocale } from "next-intl/server";
