+++ services/craft/globals/rootPages.ts
@@ -1,6 +1,6 @@
 "server-only";
 import { graphql, useFragment } from "@/gql";
-import queryAPI from "@/lib/api/client/query";
+import queryAPI from "@/lib/api/client/server";
 import { getSiteFromLocale } from "@/lib/helpers/site";
 import { getLocale } from "next-intl/server";
